refactor: 컴포넌트 분리
- SearchResult Organism으로 분리
- LectureInfo Molecule로 분리
