+++ client/src/components/UI/molecules/LectureInfo/LectureInfo.tsx
