+++ client/src/components/UI/molecules/index.ts
@@ -4,7 +4,8 @@ export type { TimetableProps } from './Timetable/Timetable';
 export { Notice } from './Notice/Notice';
 export { SelectTab } from './Timetable/SelectTab/SelectTab';
 export { SearchBar } from './SearchBar/SearchBar';
-export { SearchResults } from './SearchResults/SearchResults';
+export { LectureInfo } from './LectureInfo/LectureInfo';
+export type { LectureInfoProps, LectureInfos } from './LectureInfo/LectureInfo';
 export { ModalPopupArea } from './ModalPopupArea/ModalPopupArea';
 export type { ModalPopupAreaProps } from './ModalPopupArea/ModalPopupArea';
 export { TimeTableModalContent, TimeTableModalType } from './TimeTableModalContent/TimeTableModalContent';
