+++ client/src/components/UI/organisms/index.ts
@@ -1,1 +1,2 @@
 export { ModalPopup } from './ModalPopup/ModalPopup';
+export { SearchResults } from './SearchResults/SearchResults';
